+++ packages/isar_plus_flutter_libs/darwin/isar_plus_flutter_libs/Core/dummy.c
@@ -1,2 +1,0 @@
-// Dummy file to satisfy SPM build
-void isar_plus_dummy() {}