+++ packages/isar_plus_flutter_libs/darwin/isar_plus_flutter_libs/Core/include/isar_plus_core.h
@@ -1,0 +1,12 @@
+#pragma once
+
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+const char *isar_plus_version(void);
+
+#ifdef __cplusplus
+}
+#endif