+++ packages/isar_plus_flutter_libs/darwin/isar_plus_flutter_libs/Core/isar_plus_core.c
@@ -1,0 +1,1 @@
+#include "include/isar_plus_core.h"