+++ packages/isar_plus_flutter_libs/darwin/isar_plus_flutter_libs/Plugin/IsarPlusFlutterLibsPlugin.swift
@@ -8,15 +8,15 @@ import Cocoa
 import CIsarCore
 
 public class IsarPlusFlutterLibsPlugin: NSObject, FlutterPlugin {
+    public private(set) static var coreVersion: String?
+
     public static func register(with registrar: FlutterPluginRegistrar) {
-        let plugin = IsarPlusFlutterLibsPlugin()
-        plugin.enforceCoreBundling()
+        if let version = isar_plus_version() {
+            coreVersion = String(cString: version)
+        }
     }
 
     public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
         result(FlutterMethodNotImplemented)
-    }
-    private func enforceCoreBundling() {
-        isar_plus_force_link_all_symbols()
     }
 }