+++ tool/generate_bindings.sh
@@ -17,4 +17,3 @@ dart format --fix lib/src/native/bindings.dart
 dart format  --fix lib/src/web/bindings.dart
 
 cd ../..
-bash tool/generate_force_link_symbols.sh